pv-report-agent/requirements.txt
pandas==2.2.0
numpy==1.26.3
python-docx==1.1.0
openpyxl==3.1.2
streamlit==1.32.0
pytest==8.0.0
